add my boost any

+++ ConsoleApplication17/functions.h
@@ -198,7 +198,7 @@ void show_vector(vector_any& la)
 			cout << boost::any_cast<uint8_t>(*iter) << endl;
 
 		else if (anyone.type() == typeid(int16_t))
-			cout << boost::any_cast<int8_t>(*iter) << endl;
+			cout << boost::any_cast<int16_t>(*iter) << endl;
 
 		else if (anyone.type() == typeid(uint16_t))
 			cout << boost::any_cast<uint16_t>(*iter) << endl;
